Move: move some robotgo code to robotgo_pub.go
