+++ libei.go
@@ -35,28 +35,10 @@ import (
 	lb "github.com/go-vgo/robotgo/libei"
 )
 
-// Version of the active (libei) backend.
-const Version = lb.Version
-
-// Types — aliased so robotgo.Point == libei.Point, etc.
-type (
-	Point = lb.Point
-	Size  = lb.Size
-	Rect  = lb.Rect
-)
-
-// Bitmap mirrors the plain image descriptor used by the portable img.go
-// helpers (RGBAToBitmap, ImgToBitmap, ToRGBAGo). The Cgo backend defines an
-// identical struct in robotgo.go; under the libei tag that file is excluded,
-// so the type is provided here instead.
-type Bitmap struct {
-	ImgBuf        *uint8
-	Width, Height int
-
-	Bytewidth     int
-	BitsPixel     uint8
-	BytesPerPixel uint8
-}
+// The shared API surface — Version, GetVersion, Sleep, MilliSleep, the
+// Bitmap/Point/Size/Rect types and the DisplayID/NotPid/Scale tunables — lives
+// in the build-tag-free robotgo_pub.go and is compiled for every backend, so
+// it is NOT re-declared here.
 
 // Sentinel errors (values — aliased so errors.Is works across packages).
 var (
@@ -65,15 +47,6 @@ var (
 )
 
 // --- General ---
-
-// GetVersion get the robotgo version.
-func GetVersion() string { return lb.GetVersion() }
-
-// Sleep time.Sleep tm second.
-func Sleep(tm int) { lb.Sleep(tm) }
-
-// MilliSleep sleep tm milli second.
-func MilliSleep(tm int) { lb.MilliSleep(tm) }
 
 // Close the libei/portal connection and release resources.
 func Close() { lb.Close() }
@@ -163,7 +136,10 @@ func GetScreenSize() (int, int) { return lb.GetScreenSize() }
 func GetScaleSize(displayId ...int) (int, int) { return lb.GetScaleSize(displayId...) }
 
 // GetScreenRect get the screen rect (x, y, w, h).
-func GetScreenRect(displayId ...int) Rect { return lb.GetScreenRect(displayId...) }
+func GetScreenRect(displayId ...int) Rect {
+	r := lb.GetScreenRect(displayId...)
+	return Rect{Point{r.X, r.Y}, Size{r.W, r.H}}
+}
 
 // DisplaysNum get the number of displays.
 func DisplaysNum() int { return lb.DisplaysNum() }
